% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{clinical_data}
\alias{clinical_data}
\title{Dataset of patients with liver cancer}
\format{A data frame with 360 rows and 15 variables
\describe{
  \item{years_to_birth}{Year of birth}
  \item{Tumor_purity}{Tumor purity}
  \item{pathologic_stage}{Pathologic stage of tumor}
  \item{pathology_T_stage}{Pathologic stage of tumor}
  \item{pathology_N_stage}{Pathologic stage of tumor}
  \item{pathology_M_stage}{Pathologic stage of tumor}
  \item{histological_type}{Histological type of tumor}
  \item{gender}{Gender of patient}
  \item{radiation_therapy}{Radiation therapy (yes or no)}
  \item{residual_tumor}{Residual tumor}
  \item{race}{Race}
  \item{ethnicity}{Ethnicity}
  \item{overallsurvival}{Survival time}
  \item{status}{Death status}
  \item{overall_survival}{Survival time (rounded)}
}}
\source{
\url{http://www.linkedomics.org/}
}
\description{
A data frame containing clinical data for liver cancer dataset
}
