% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{liver}
\alias{liver}
\title{Dataset of patients with liver cancer}
\format{A list of data frame with 360 rows
\describe{
  \item{meth}{Methylation data}
  \item{mutation}{Mutation data}
  \item{RNA}{Gene expression data}
  \item{CNV}{Copy number variation data}
}}
\source{
\url{http://www.linkedomics.org/}
}
\description{
A data frame containing genomic data for liver cancer dataset
}
