% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{obesity}
\alias{obesity}
\title{Dataset of obsese individuals}
\format{A list of data frame with 26 rows
\describe{
  \item{meth}{Methylation data}
  \item{mRNA}{Gene expression data}
  \item{miRNA}{miRNA data}
}}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE58248}
}
\description{
A data frame containing genomic data for obesity dataset
}
